backend/requirements.txt
flask
flask-cors
Flask-Migrate
Flask-SQLAlchemy
gunicorn
psycopg2-binary
